refine test for stacked dense unet
